requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
markovify==0.9.0
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.15.0
Unidecode==1.2.0
Werkzeug==1.0.1
